+++ clients/admin-ui/src/features/system/system-form-declaration-tab/PrivacyDeclarationForm.tsx
@@ -35,7 +35,6 @@ import {
   PrivacyDeclarationResponse,
   ResourceTypes,
 } from "~/types/api";
-import { Cookies } from "~/types/api/models/Cookies";
 
 export const ValidationSchema = Yup.object().shape({
   data_categories: Yup.array(Yup.string())
@@ -46,7 +45,6 @@ export const ValidationSchema = Yup.object().shape({
 
 export type FormValues = Omit<PrivacyDeclarationResponse, "cookies"> & {
   customFieldValues: CustomFieldValues;
-  cookies?: string[];
 };
 
 const defaultInitialValues: FormValues = {
@@ -67,16 +65,10 @@ const defaultInitialValues: FormValues = {
   third_parties: "",
   shared_categories: [],
   customFieldValues: {},
-  cookies: [],
   id: "",
 };
 
 const transformFormValueToDeclaration = (values: FormValues) => {
-  // transform cookies from strings into object with default values
-  const transformedCookies = values.cookies
-    ? values.cookies.map((name) => ({ name, path: "/" }))
-    : undefined;
-
   const declaration = {
     ...values,
     // fill in an empty string for name: https://github.com/ethyca/fideslang/issues/98
@@ -90,7 +82,6 @@ const transformFormValueToDeclaration = (values: FormValues) => {
     shared_categories: values.data_shared_with_third_parties
       ? values.shared_categories
       : undefined,
-    cookies: transformedCookies,
   };
   return declaration;
 };
@@ -101,7 +92,6 @@ export interface DataProps {
   allDataSubjects: DataSubject[];
   allDatasets?: Dataset[];
   includeCustomFields?: boolean;
-  cookies?: Cookies[] | null;
 }
 
 export const PrivacyDeclarationFormComponents = ({
@@ -109,7 +99,6 @@ export const PrivacyDeclarationFormComponents = ({
   allDataCategories,
   allDataSubjects,
   allDatasets,
-  cookies,
   values,
   includeCustomFields,
   privacyDeclarationId,
@@ -323,21 +312,6 @@ export const PrivacyDeclarationFormComponents = ({
           </Collapse>
         </Stack>
       </SystemFormInputGroup>
-      <SystemFormInputGroup heading="Cookies">
-        <ControlledSelect
-          name="cookies"
-          label="Cookies"
-          options={
-            cookies && cookies.length
-              ? cookies.map((c) => ({ label: c.name, value: c.name }))
-              : []
-          }
-          mode="tags"
-          tooltip="Which cookies are placed on consumer domains for this purpose?"
-          layout="stacked"
-          disabled={lockedForGVL}
-        />
-      </SystemFormInputGroup>
       {includeCustomFields ? (
         <CustomFieldsList
           resourceType={ResourceTypes.PRIVACY_DECLARATION}
@@ -352,18 +326,12 @@ export const transformPrivacyDeclarationToFormValues = (
   privacyDeclaration?: PrivacyDeclarationResponse,
   customFieldValues?: CustomFieldValues,
 ): FormValues => {
-  if (privacyDeclaration) {
-    const formCookies =
-      privacyDeclaration.cookies && privacyDeclaration.cookies.length > 0
-        ? privacyDeclaration.cookies.map((c) => c.name)
-        : undefined;
-    return {
-      ...privacyDeclaration,
-      customFieldValues: customFieldValues || {},
-      cookies: formCookies,
-    };
-  }
-  return defaultInitialValues;
+  return privacyDeclaration
+    ? {
+        ...privacyDeclaration,
+        customFieldValues: customFieldValues || {},
+      }
+    : defaultInitialValues;
 };
 
 /**